+++ app/components/url/url-controller.js
@@ -14,6 +14,7 @@
 angular.module('lizard-nxt')
 .controller('UrlController', [
   '$scope',
+  '$q',
   '$timeout',
   'LocationGetterSetter',
   'UrlState',
@@ -26,8 +27,10 @@ angular.module('lizard-nxt')
   'State',
   'LeafletService',
   'gettextCatalog',
+  'FavouritesService',
   function (
     $scope,
+    $q,
     $timeout,
     LocationGetterSetter,
     UrlState,
@@ -39,7 +42,8 @@ angular.module('lizard-nxt')
     NxtRegionsLayer,
     State,
     LeafletService,
-    gettextCatalog
+    gettextCatalog,
+    FavouritesService
   ) {
 
     // Configuration object for url state.
@@ -219,7 +223,7 @@ angular.module('lizard-nxt')
         state.boxType.part, state.boxType.index, State.box.type
       );
 
-      if (old === 'point' || old === 'line' || old === 'region' || old === 'multi-point') {
+      if (['point', 'line', 'region', 'multi-point'].indexOf(old) != -1) {
         // Remove geometry from url
         LocationGetterSetter.setUrlValue(
           state.geom.part, state.geom.index, undefined);
@@ -245,17 +249,47 @@ angular.module('lizard-nxt')
     });
 
     /**
+     * @function
+     * @description Checks if the url is a `favourite url`. And then proceeds to
+     * fetch the favourites as it is asked.
+     * return {object} - thennable promise which resolves to true/false
+     */
+    var favouritesFromUrl = function () {
+      var firstUrlPart = LocationGetterSetter.getUrlValue(
+        'path', 0);
+      if (firstUrlPart !== 'favourites') {
+        return false;
+      } else {
+        var favouriteUUID = LocationGetterSetter.getUrlValue(
+          'path',
+          1
+        );
+        return favouriteUUID;
+      }
+    };
+
+    /**
      * Set the state from the url on init or set the url from the default state
      * when the url is empty.
      */
-    var setStateFromUrl = function () {
-      var language = LocationGetterSetter.getUrlValue(state.language.part, state.language.index),
-        boxType = LocationGetterSetter.getUrlValue(state.boxType.part, state.boxType.index),
-        geom = LocationGetterSetter.getUrlValue(state.geom.part, state.geom.index),
-        layerGroupsFromURL = LocationGetterSetter.getUrlValue(state.layerGroups.part, state.layerGroups.index),
-        mapView = LocationGetterSetter.getUrlValue(state.mapView.part, state.mapView.index),
-        time = LocationGetterSetter.getUrlValue(state.timeState.part, state.timeState.index),
-        context = LocationGetterSetter.getUrlValue(state.context.part, state.context.index);
+    var setStateFromUrl = function (favouriteURL) {
+
+      if (!favouriteURL) {
+        var language = LocationGetterSetter.getUrlValue(
+            state.language.part, state.language.index),
+          boxType = LocationGetterSetter.getUrlValue(
+            state.boxType.part, state.boxType.index),
+          geom = LocationGetterSetter.getUrlValue(
+            state.geom.part, state.geom.index),
+          layerGroupsFromURL = LocationGetterSetter.getUrlValue(
+            state.layerGroups.part, state.layerGroups.index),
+          mapView = LocationGetterSetter.getUrlValue(
+            state.mapView.part, state.mapView.index),
+          time = LocationGetterSetter.getUrlValue(
+            state.timeState.part, state.timeState.index),
+          context = LocationGetterSetter.getUrlValue(
+            state.context.part, state.context.index);
+      }
 
       setLanguage(language);
 
@@ -267,15 +301,17 @@ angular.module('lizard-nxt')
 
       if (context) {
         $scope.transitionToContext(context);
-      } else {
-        LocationGetterSetter.setUrlValue(state.context.part, state.context.index, state.context.value);
+      } else if (!favouriteURL) {
+        LocationGetterSetter.setUrlValue(
+          state.context.part, state.context.index, state.context.value);
         $scope.transitionToContext(state.context.value);
       }
 
       if (boxType) {
         State.box.type = boxType;
       } else {
-        LocationGetterSetter.setUrlValue(state.boxType.part, state.boxType.index, State.box.type);
+        LocationGetterSetter.setUrlValue(
+          state.boxType.part, state.boxType.index, State.box.type);
       }
 
       if (geom) {
@@ -304,7 +340,21 @@ angular.module('lizard-nxt')
 
     };
 
-    setStateFromUrl();
+
+    var favouriteUUID = favouritesFromUrl();
+    if (favouriteUUID) {
+      var deferred = $q.defer();
+      FavouritesService.getFavourite(
+        favouriteUUID,
+        function (favourite, getResponseHeaders) {
+          FavouritesService.applyFavourite(favourite);
+          deferred.resolve(true);
+        });
+      deferred.promise.then(setStateFromUrl);
+    }
+    else {
+      setStateFromUrl(false);
+    }
 
   }
 ]);