+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog of lizard-nxt client
 Unreleased (2.9.0) (XXXX-XX-XX)
 -------------------------------
 
+- Add share favourites.
+
 - Remove layergroups from search.
 
 - Adapt search results to updated full-text search API response.